+++ odoo/local-src/sf_mrp/wizard/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# Part of Sensefly.
+import mrp_product_produce