+++ odoo/local-src/sf_account_payment_order/__manifest__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'SenseFly Account Payment Order',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'category': 'SenseFly',
+    'depends': ['account_payment_order'],
+    'data': [],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'auto_install': False,
+}